Fix a bug in EIC irq handler
as an ANA 2nd level irq, which is also a cascading irq source,
eic_handler should do the chip->unmask explicitly. Missing of
doing this can cause problems like missing power key irq,
which is a serious "unable to wake up the phone" trouble for the
end user.
All the related workaround code is removed too.

	modified:   arch/arm/mach-sc8810/eic.c
	modified:   arch/arm/mach-sc8810/gpio.c
	modified:   arch/arm/mach-sc8810/pm_vlx.c
	modified:   arch/arm/mach-sc8810/timer.c
	modified:   drivers/input/touchscreen/tp_sprd.c

Change-Id: I12cedd86bb0395b7f88cff77eea3e30194779aef