Add driver for the h/w watchdog
This is a hw watchdog starts after init, and fed by a timer started
by the watchdog driver itself. So it could be treat as a watchdog
for the timer subsystem.
This watchdog can also be fed from userland. It will be stopped
when system suspends, and restarted when system resumes.

Change-Id: I238090fcf247892e6cd6229dbf6a7bdec95dd20e