bug#000 fix a hang related to OOM
In some cases, some kernel thread (such as sync_supers) locks
some memory zone in out of memory, if there are no killable processes.

Subsequently, the process on user space has been resumed from
the uninterruptible state, and continues to request memory.
it would be blocked, and try to allocate memory again after a short
interval.

However, if there is no any killable process except the caller itself,
it should be a deadlock!

This breaks the deadlock to allow the caller process killable.

Change-Id: Ib668a680948941f08bbc75f14b1c548b9482e2e6
Signed-off-by: Horace Hsieh <horace.xie@spreadtrum.com>