bug#255480 pmem: share spare reserved pages as page cache
When reading in file for a page fault or reading ahead pages,
This tries to allocate pages from reserved pmem pool,
if the vma is for the share object (read-only and executable).

The shared reservation pages are not added into any
existing LRU lists, but be kept into a self-own list,
which is also shrinkable like others.

Moreover, if some devices require more physical memory,
it would take a higher priority to get the pages that
have been used as page cache.

Change-Id: Ic8cd435f169beb19186d6f1caa299386afb7cef9
Signed-off-by: Horace Hsieh <horace.xie@spreadtrum.com>
Signed-off-by: Yanbin Fang <yanbin.fang@spreadtrum.com>