bug#266397 oom-kill: grant TIF_MEMDIE to all thread group members
In some situation, a thread group member holds the mm->mmap_sem
as a writer and then blocks waiting for memory.

After that, the oom-killer chooses this particular thread group
to die to release some memory. since the mm->mmap_sem is held,
the thread group cannot finish the exit routine,
therefore no memory can be released, which leads to a deadlock
situation.

This grants TIF_MEMDIE to all thread group members
to avoid the deadlock.

Change-Id: I9efb9c4e65157bc04568aadfa93cc85aca02b2ab
Signed-off-by: Horace Hsieh <horace.xie@spreadtrum.com>
Signed-off-by: Ren Geng <geng.ren@spreadtrum.com>