bug#278235 mm: page_alloc: avoid direct reclaim for high-order allocations
When the compaction of memory was disabled, and system went
into the fragmented state, the high-order allocation could
not be satisfied from freelist, and then it would try to
perform direct reclaim, no matter if NO_KSWAPD was marked.

Such actions heavily impacted the performance,
and were not what we expected, so this avoids them.

Change-Id: Id4264fc40016cd4bee3c0b3748362069851ceff0
Signed-off-by: Horace Hsieh <horace.xie@spreadtrum.com>