Bug #320122 ion: system_heap: optimize cached allocation
Occasionally, we found it cost around 20-30% of cpu in
allocating the cached allocaton from ion_system_heap,
when playing subway surfer, which significantly
impacted the system performance.

This improves the process of allocation for cached buffer
in ion_system_heap to implement the real deferred release,
and further avoids direct allocating from buddy
system if there are the defer-released pages in the list.

Also, the ion_system_heap_shrink is extended to shrink
these pages in the new list.

[ self test  ] tested on 7730ec

Change-Id: Iedd8647d2ec97b6b658edb88cc1d1707f18465da
Signed-off-by: Horace Hsieh <horace.xie@spreadtrum.com>